sdl2: SDL_CondWait patch
FAKE_RECURSIVE_MUTEX is turned on on Genode because our pthread_mutex*
implementation is not resursive. On SDL_CondWait a phtread_mutex is handed in
locked state and is unlocked in pthread_cond_wait and locked at the end of this
function. If in meantime another thread uses the same SDL mutex and frees it,
the SDL owner will be set to 0. At the end of SDL_CondWait we therefore need to
set the owner in case it is null since we know we own the phtread_mutex.