SDL2: preliminary window resizing support
This commit introduces preliminary support for resizing a window.
How and if it works depends to some degree on the program using
the SDL2 library itself. It works well for the currently tested
ports like chocolate-doom and yquake2 (software renderer).
On the other hand program using an OpenGL context via SDL2 still
have issues resizing the window (EGL errors).

As an additional feature the commit adds support for setting the
initial window geometry via the '<initial>' node, e.g.:

! <initial width="800" height="600"/>

This prevents overly large windows because the actual screen size
is propagated by nitpicker.