Fix an issue with ATH9K under the new packet-queueing implementation.
For incoming urbs, we need to increment available_work when a space in the queue opens. FOr outgoing this is unnecessary, so it caused the available_work variable to exceed the amount we actually had avaiable, hence the loop ran forever. (This was apparently only an issue for the bulk endpoint, since outgoing interrupt transfers are used one-at-a-time.)